refactor: move IMA operator into internal/attest package
